requirements.txt
asgiref==3.3.4
awscli==1.19.24
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
botocore==1.20.24
certifi==2021.5.30
cffi==1.14.6
click==8.0.1
colorama==0.4.3
distlib==0.3.2
djangorestframework==3.12.4
dnspython==2.1.0
docutils==0.15.2
email-validator==1.1.3
fastapi==0.68.0
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
numpy==1.19.5
pipenv==2021.5.29
platformdirs==2.0.2
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.3.4
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
starlette==0.14.2
typing-extensions==3.10.0.0
virtualenv==20.6.0
virtualenv-clone==0.5.5
Werkzeug==2.0.1
wincertstore==0.2
WTForms==2.3.3
